fix(api): rename internal api connection->connections
